Refactoring phase 3
Update route path

Thêm ResponseUtil để chuẩn hóa phản hồi API và chuyển đổi các hằng số thông báo sang ResultCode
- Tạo lớp ResponseUtil để tạo phản hồi API thống nhất
- Chuyển đổi các hằng số MessageConstant và MessageCodeConstant sang enum ResultCode
- Cập nhật các controller để sử dụng ResponseUtil và ResultCode
- Loại bỏ các hằng số thông báo cũ và các phương thức tạo phản hồi trực tiếp trong controller

Migrate MessageCodeConstant & MessageConstant to ResultCode
Add pagination and entity graph support to UserRepository
Normalize email to lowercase during user registration in AuthController
Enhance BlogPostResponse to include user details and update BlogPostRepository queries to join with User for additional information; implement access control in deletePost method of BlogPostServiceImpl
Add success messages for create, update, and delete operations in controllers
Add unit tests for CollectionService methods
Refactor ApiExceptionHandler to use exception messages directly in response
Update UserController to allow authenticated users access and add emailVerified field to UserResponse
Refactor FreePatternMapper and PatternMapper to use toEntities method for file mapping
Refactor FileMapper and model classes to use Set instead of List for improved performance and uniqueness
Refactor model relationships to use Set instead of List for improved performance and uniqueness

Add MyBatis integration for collection and free pattern management
- Introduced MyBatis dependencies in build.gradle
- Implemented ColMapper for querying collections by user and free pattern
- Enhanced ColFrepRepo with additional queries for collection and free pattern relationships
- Created CollectionAvatarService and its implementation for managing collection avatars
- Updated CollectionServiceImpl to utilize the new avatar service and streamline pattern addition

Refactor response handling in controllers to standardize success messages and update pagination parameters
Remove 'files' from EntityGraph attributePaths in FreePatternRepository
Replace tsid-creator import with hypersistence-utils-hibernate and remove id initialization method in BaseEntity
Replace tsid-creator dependency with hypersistence-utils-hibernate-63 in build.gradle
Refactor CategoryServiceImpl to use Optional for findById and add @Transactional annotation
Add pageable query for FreePattern by user and refactor response handling
Use AppConstant for date pattern in UserProfileRequest
Add DatasourceConfig class and migrate HikariCP settings from application.yml
Refactor pagination parameters in BlogController and BlogPostService

+++ src/main/java/org/crochet/service/impl/BlogPostServiceImpl.java
@@ -1,7 +1,9 @@
 package org.crochet.service.impl;
 
 import lombok.extern.slf4j.Slf4j;
-import org.crochet.constant.MessageConstant;
+import org.crochet.enums.ResultCode;
+import org.crochet.enums.RoleType;
+import org.crochet.exception.AccessDeniedException;
 import org.crochet.exception.ResourceNotFoundException;
 import org.crochet.mapper.BlogPostMapper;
 import org.crochet.mapper.FileMapper;
@@ -18,6 +20,7 @@ import org.crochet.repository.BlogPostRepository;
 import org.crochet.repository.GenericFilter;
 import org.crochet.service.BlogPostService;
 import org.crochet.util.ImageUtils;
+import org.crochet.util.SecurityUtils;
 import org.crochet.util.SettingsUtil;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
@@ -30,8 +33,7 @@ import org.springframework.util.StringUtils;
 
 import java.util.List;
 import java.util.Map;
-
-import static org.crochet.constant.MessageCodeConstant.MAP_CODE;
+import java.util.Objects;
 
 /**
  * BlogPostServiceImpl class
@@ -70,8 +72,8 @@ public class BlogPostServiceImpl implements BlogPostService {
             BlogCategory blogCategory = null;
             if (request.getBlogCategoryId() != null) {
                 blogCategory = blogCategoryRepo.findById(request.getBlogCategoryId())
-                        .orElseThrow(() -> new ResourceNotFoundException(MessageConstant.MSG_BLOG_CATEGORY_NOT_FOUND,
-                                MAP_CODE.get(MessageConstant.MSG_BLOG_CATEGORY_NOT_FOUND)));
+                        .orElseThrow(() -> new ResourceNotFoundException(ResultCode.MSG_BLOG_CATEGORY_NOT_FOUND.message(),
+                                ResultCode.MSG_BLOG_CATEGORY_NOT_FOUND.code()));
             }
             var images = ImageUtils.sortFiles(request.getFiles());
             blogPost = BlogPost.builder()
@@ -91,18 +93,18 @@ public class BlogPostServiceImpl implements BlogPostService {
     /**
      * Retrieves a paginated list of blog posts based on the provided parameters.
      *
-     * @param pageNo   The page number to retrieve (0-indexed).
-     * @param pageSize The number of blog posts to include in each page.
-     * @param sortBy   The attribute by which the blog posts should be sorted.
-     * @param sortDir  The sorting direction, either "ASC" (ascending) or "DESC"
-     *                 (descending).
-     * @param filters  The list of filters.
+     * @param offset  The page number to retrieve (0-indexed).
+     * @param limit   The number of blog posts to include in each page.
+     * @param sortBy  The attribute by which the blog posts should be sorted.
+     * @param sortDir The sorting direction, either "ASC" (ascending) or "DESC"
+     *                (descending).
+     * @param filters The list of filters.
      * @return A {@link org.crochet.payload.response.PaginationResponse} containing the paginated list of
      * blog posts.
      */
     @Override
-    public PaginationResponse<BlogPostResponse> getBlogs(int pageNo, int pageSize, String sortBy, String sortDir,
-                                       Filter[] filters) {
+    public PaginationResponse<BlogPostResponse> getBlogs(int offset, int limit, String sortBy, String sortDir,
+                                                         Filter[] filters) {
         Specification<BlogPost> spec = Specification.where(null);
         if (filters != null && filters.length > 0) {
             GenericFilter<BlogPost> filter = GenericFilter.create(filters);
@@ -116,7 +118,7 @@ public class BlogPostServiceImpl implements BlogPostService {
 
         Sort.Direction dir = Sort.Direction.fromString(sortDir);
         Sort sort = Sort.by(dir, sortBy);
-        Pageable pageable = PageRequest.of(pageNo, pageSize, sort);
+        Pageable pageable = PageRequest.of(offset, limit, sort);
         var page = blogPostRepo.findBlogOnHomeWithIds(blogIds, pageable);
 
         return PaginationMapper.getInstance().toPagination(page);
@@ -125,13 +127,13 @@ public class BlogPostServiceImpl implements BlogPostService {
     /**
      * Get blog ids
      *
-     * @param pageNo Page number
+     * @param offset Page number
      * @param limit  Limit
      * @return List of blog ids
      */
     @Override
-    public List<String> getBlogIds(int pageNo, int limit) {
-        Pageable pageable = PageRequest.of(pageNo, limit);
+    public List<String> getBlogIds(int offset, int limit) {
+        Pageable pageable = PageRequest.of(offset, limit);
         return blogPostRepo.getBlogIds(pageable);
     }
 
@@ -148,8 +150,8 @@ public class BlogPostServiceImpl implements BlogPostService {
     @Override
     public BlogPostResponse getDetail(String id) {
         var blogPost = blogPostRepo.getDetail(id).orElseThrow(
-                () -> new ResourceNotFoundException(MessageConstant.MSG_BLOG_NOT_FOUND,
-                        MAP_CODE.get(MessageConstant.MSG_BLOG_NOT_FOUND)));
+                () -> new ResourceNotFoundException(ResultCode.MSG_BLOG_NOT_FOUND.message(),
+                        ResultCode.MSG_BLOG_NOT_FOUND.code()));
         return BlogPostMapper.INSTANCE.toResponse(blogPost);
     }
 
@@ -188,7 +190,16 @@ public class BlogPostServiceImpl implements BlogPostService {
     @Transactional
     @Override
     public void deletePost(String id) {
+        var currentUser = SecurityUtils.getCurrentUser();
+        if (currentUser == null) {
+            throw new ResourceNotFoundException(ResultCode.MSG_USER_NOT_FOUND.message(),
+                    ResultCode.MSG_USER_NOT_FOUND.code());
+        }
         var blogPost = getById(id);
+        var isAdmin = Objects.equals(currentUser.getRole(), RoleType.ADMIN);
+        if (!isAdmin && !Objects.equals(currentUser.getId(), blogPost.getCreatedBy())) {
+            throw new AccessDeniedException("Cannot delete this post");
+        }
         blogPostRepo.delete(blogPost);
     }
 
@@ -202,7 +213,7 @@ public class BlogPostServiceImpl implements BlogPostService {
      */
     private BlogPost getById(String id) {
         return blogPostRepo.findById(id).orElseThrow(
-                () -> new ResourceNotFoundException(MessageConstant.MSG_BLOG_NOT_FOUND,
-                        MAP_CODE.get(MessageConstant.MSG_BLOG_NOT_FOUND)));
+                () -> new ResourceNotFoundException(ResultCode.MSG_BLOG_NOT_FOUND.message(),
+                        ResultCode.MSG_BLOG_NOT_FOUND.code()));
     }
 }
